+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -119,5 +119,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-07T12:04:05.293Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-07T14:39:06.330Z")
     r"""The date and time the resource was last updated"""