+++ packages/orq-rc/RELEASES.md
@@ -2189,3 +2189,13 @@ Based on:
 - [python v3.11.0-rc.1] packages/orq-rc
 ### Releases
 - [PyPI v3.11.0-rc.1] https://pypi.org/project/orq-ai-sdk/3.11.0-rc.1 - packages/orq-rc
+
+## 2025-08-07 14:40:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.599.0 (2.675.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.11.0-rc.2] packages/orq-rc
+### Releases
+- [PyPI v3.11.0-rc.2] https://pypi.org/project/orq-ai-sdk/3.11.0-rc.2 - packages/orq-rc